requirements.txt
streamlit==1.27.1
plotly==5.17.0
pandas==2.1.1
numpy==1.25.2
networkx==3.1
Pillow==10.0.1
scipy==1.11.3
statsmodels==0.14.0
lifelines==0.27.8
arviz==0.20.0
pymc==5.20.0
scikit-learn==1.3.1
matplotlib==3.8.0
